@@ -2,7 +2,7 @@ package com.trinitymirror.reuseadvertssample
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.text.Editable
 import android.text.TextWatcher
 import com.trinitymirror.reuseadvertssample.adverts.AdvertsContainerActivity
