@@ -1,7 +1,7 @@
 package com.trinitymirror.reuseadvertssample.adverts
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -9,8 +9,8 @@ import android.widget.Toast
 import com.google.android.gms.ads.doubleclick.PublisherAdRequest
 import com.google.android.gms.ads.formats.UnifiedNativeAd
 import com.google.android.gms.ads.formats.UnifiedNativeAdView
-import com.mirror.news.ui.adapter.teaser_list.holder.TeaserAppInstallAdViewHolder
-import com.mirror.news.ui.adapter.teaser_list.holder.TeaserContentAdViewHolder
+import com.trinitymirror.reuseadvertssample.adverts.holder.TeaserAppInstallAdViewHolder
+import com.trinitymirror.reuseadvertssample.adverts.holder.TeaserContentAdViewHolder
 import com.trinitymirror.reuseadvertssample.MyApplication
 import com.trinitymirror.reuseadvertssample.R
 import com.trinitymirror.reuseadvertssample.util.DfpNativeAdRequest
