@@ -1,9 +1,9 @@
 package com.trinitymirror.reuseadvertssample.adverts
 
 import android.os.Bundle
-import android.support.design.widget.BottomNavigationView
-import android.support.v4.app.Fragment
-import android.support.v7.app.AppCompatActivity
+import com.google.android.material.bottomnavigation.BottomNavigationView
+import androidx.fragment.app.Fragment
+import androidx.appcompat.app.AppCompatActivity
 import com.trinitymirror.reuseadvertssample.R
 import kotlinx.android.synthetic.main.activity_adverts.*
 
