@@ -1,6 +1,6 @@
-package com.mirror.news.ui.adapter.teaser_list.holder
+package com.trinitymirror.reuseadvertssample.adverts.holder
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.widget.Button
 import android.widget.TextView
@@ -8,7 +8,6 @@ import com.google.android.gms.ads.formats.MediaView
 import com.google.android.gms.ads.formats.UnifiedNativeAd
 import com.google.android.gms.ads.formats.UnifiedNativeAdView
 import com.trinitymirror.reuseadvertssample.R
-import io.reactivex.subjects.PublishSubject
 
 class TeaserContentAdViewHolder(view: View) : RecyclerView.ViewHolder(view) {
 
